+++ presentation/src/main/java/com/dcs/presentation/ui/trend/MovieItem.kt
@@ -22,7 +22,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.bumptech.glide.integration.compose.ExperimentalGlideComposeApi
 import com.bumptech.glide.integration.compose.GlideImage
-import com.dcs.presentation.common.CircularProgressBarWithPercentage
+import com.dcs.presentation.core.common.CircularProgressBarWithPercentage
 import com.dcs.presentation.core.extensions.toImageUrl
 import com.dcs.presentation.core.model.MovieItemUiState
